requirements.txt
# Web框架和核心扩展
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.1
Flask-Migrate>=4.0.0
Flask-Bootstrap>=3.3.7
Flask-Moment>=1.0.2
Flask-APScheduler
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Limiter>=3.5.0

# 数据处理
pandas>=1.3.0
openpyxl>=3.0.7

# OCR服务
tencentcloud-sdk-python

# 安全与认证
bcrypt>=4.0.0
pyotp>=2.9.0
qrcode>=7.4

# 工具
python-dotenv>=0.19.0

# 部署工具
gunicorn>=20.1.0
